requirements.txt
ultralytics==8.1.43
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.1
torch==2.2.2
torchvision==0.17.2
PyYAML==6.0.1
tqdm==4.66.2
pyqt5==5.15.11
openai-whisper
